refactor: use several executors instead mixins
