+++ ansible/executors/__init__.py
@@ -1,0 +1,6 @@
+"""
+Package contains different ansible-runner executors
+"""
+from ansible.executors.abstract_executor import AbstractAnsibleExecutor
+from ansible.executors.module_executor import AnsibleModuleExecutor
+from ansible.executors.playbook_executor import AnsiblePlaybookExecutor
