--- ansible/mixins/__init__.py
@@ -1,5 +1,0 @@
-"""
-Contains mixins with ansible execution methods
-"""
-from ansible.mixins.module_executor import ModuleExecutorMixin
-from ansible.mixins.playbook_executor import PlaybookExecutorMixin
